feat: add missing files or updates from dot folders
