+++ .nvmrc
@@ -1,1 +1,1 @@
-16.19.0
+18.17.1
