--- .storybook/preview.ts
@@ -1,11 +1,0 @@
-export const parameters = {
-  actions: { argTypesRegex: '^on[A-Z].*' },
-  controls: {
-    matchers: {
-      color: /(background|color)$/i,
-      date: /Date$/,
-    },
-  },
-
-  layout: 'centered',
-};
