+++ .storybook/preview.tsx
@@ -1,0 +1,27 @@
+import { initialize, mswLoader } from 'msw-storybook-addon';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+
+initialize({ onUnhandledRequest: 'bypass' });
+const queryClient = new QueryClient();
+
+export const parameters = {
+  actions: { argTypesRegex: '^on[A-Z].*' },
+  controls: {
+    matchers: {
+      color: /(background|color)$/i,
+      date: /Date$/,
+    },
+  },
+
+  layout: 'centered',
+};
+
+export const decorators = [
+  (Story) => (
+    <QueryClientProvider client={queryClient}>
+      <Story />
+    </QueryClientProvider>
+  ),
+];
+
+export const loaders = [mswLoader];
